NGAF-304 Feature/create new onboarding (#21)
* fix: change animation
feat: create custom view for time preference
feat: create custom view for days button stack

* feat: create the new onboarding page

* fix: make minimum one day active

* fix: UI Constraints

* fix: simplify date picker handling

* fix: max date, add value changed handler, change delegate property name

* fix: date inital value

* fix: UI constraint refactor

* fix: indentation

* fix!: rename variable name

* fix: simplify reminder handler, change variable name

* fix!: move date picker component to itself

* fix: adjust based on design

* fix: adjust based on design and fix day list not hidden after unchecking different hour

* fix!: change user manager to singleton, add do catch

* fix: migrate singleton class

* fix: indentation

* feat: add variable to store data

* feat: add weekday enum

* fix: remove isEnabled prop

* fix: identation

* feat: custom hasher

* fix: update class singleton

* fix: class singleton

* feat: add isEnabled prop

* fix: static weekday name

* feat: add working hours, implement enable/disable day

* feat: add reset date

* feat: update day changed flow and reset after unselected

* fix: remove id, add isFlexibleWorkHour, add default value

* feat: add save preference data

* fix: change animation
feat: create custom view for time preference
feat: create custom view for days button stack

* feat: remove unused files, remove unused coredata entity

* feat: remove unused files, remove unused coredata entity

* fix!: remove unused entity

* feat: fix file references

---------

+++ climbr/Models/Character.swift
@@ -11,7 +11,7 @@ struct CharacterModel: Codable {
     var id: UUID = UUID()
     var name: String
     var gender: Gender
-    var point: Int64
+    var point: Int
     var headEquipment: EquipmentItem = .emptyHG
     var handEquipment: EquipmentItem = .emptyS
     var backEquipment: EquipmentItem = .emptyBP
